TPP-131 fix: update screens depend on new models

+++ frontend/lib/screens/calendar_screen.dart
@@ -532,7 +532,7 @@ class _CalendarScreenState extends State<CalendarScreen> {
     );
   }
 
-  Widget _buildRewardBadge(double reward) {
+  Widget _buildRewardBadge(int reward) {
     return Container(
       padding: const EdgeInsets.symmetric(horizontal: 6, vertical: 2),
       decoration: BoxDecoration(
